requirements.txt
adagio==0.2.6
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.3.2
cloudpickle==3.1.1
clr-loader==0.2.5
cmdstanpy==1.2.5
colorama==0.4.6
contourpy==1.3.1
coreforecast==0.0.15
cycler==0.12.1
distlib==0.3.6
et-xmlfile==1.1.0
filelock==3.9.0
fonttools==4.56.0
fs==2.4.16
fsspec==2025.3.0
ftfy==6.3.1
fugue==0.9.1
h11==0.14.0
holidays==0.69
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
kiwisolver==1.4.8
llvmlite==0.44.0
lxml==5.3.1
matplotlib==3.10.1
memory-profiler==0.61.0
nixtlats==0.1.14
numba==0.61.0
numpy==2.1.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.2
patsy==1.0.1
pillow==11.1.0
platformdirs==2.6.2
polars==1.25.2
prophet==1.1.6
psutil==7.0.0
pyarrow==19.0.1
pycparser==2.21
pydantic==1.10.21
pydantic_core==2.27.2
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.8.2
pythonnet==3.0.1
pytz==2024.2
regex==2024.11.6
requests==2.32.3
scipy==1.15.2
seaborn==0.13.2
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stanio==0.5.1
statsforecast==2.0.1
statsmodels==0.14.4
tenacity==9.0.0
threadpoolctl==3.6.0
tqdm==4.66.5
triad==0.9.8
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
utilsforecast==0.2.12
virtualenv==20.17.1
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
